requirements.txt
fastapi
uvicorn
requests
langgraph
langchain
langchain-community
sentence-transformers
streamlit
python-dotenv